# Refactor ReplayBuffer to to use PyTorch tensors avoiding numpy