--- a/amp_rsl_rl/storage/replay_buffer.py
+++ b/amp_rsl_rl/storage/replay_buffer.py
@@ -3,10 +3,8 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-
 import torch
-import numpy as np
-from typing import Generator, Tuple
+from typing import Generator, Tuple, Union
 
 
 class ReplayBuffer:
@@ -17,29 +15,44 @@ class ReplayBuffer:
         states (Tensor): Buffer of current states.
         next_states (Tensor): Buffer of next states.
         buffer_size (int): Maximum number of elements in the buffer.
-        device (str): Device where tensors are stored.
+        device (str or torch.device): Device where tensors are stored.
         step (int): Current write index.
         num_samples (int): Total number of inserted samples (up to buffer_size).
     """
 
-    def __init__(self, obs_dim: int, buffer_size: int, device: str) -> None:
+    def __init__(
+        self,
+        obs_dim: int,
+        buffer_size: int,
+        device: Union[str, torch.device] = "cpu",
+    ) -> None:
         """
         Initialize a ReplayBuffer object.
 
         Args:
             obs_dim (int): Dimension of the observation space.
             buffer_size (int): Maximum number of transitions to store.
-            device (str): Torch device where buffers are allocated ('cpu' or 'cuda').
+            device (str or torch.device): Torch device where buffers are allocated ('cpu' or 'cuda').
         """
-        self.states = torch.zeros(buffer_size, obs_dim).to(device)
-        self.next_states = torch.zeros(buffer_size, obs_dim).to(device)
+        self.device = torch.device(device)
         self.buffer_size = buffer_size
-        self.device = device
+
+        # Pre-allocate buffers on the target device
+        self.states = torch.zeros(
+            (buffer_size, obs_dim), dtype=torch.float32, device=self.device
+        )
+        self.next_states = torch.zeros(
+            (buffer_size, obs_dim), dtype=torch.float32, device=self.device
+        )
 
         self.step = 0
         self.num_samples = 0
 
-    def insert(self, states: torch.Tensor, next_states: torch.Tensor) -> None:
+    def insert(
+        self,
+        states: torch.Tensor,
+        next_states: torch.Tensor,
+    ) -> None:
         """
         Add a batch of states and next_states to the buffer.
 
@@ -47,26 +60,33 @@ class ReplayBuffer:
             states (Tensor): Batch of current states (batch_size, obs_dim).
             next_states (Tensor): Batch of next states (batch_size, obs_dim).
         """
-        num_states = states.shape[0]
-        start_idx = self.step
-        end_idx = self.step + num_states
+        # Move incoming data to buffer device if necessary
+        states = states.to(self.device)
+        next_states = next_states.to(self.device)
 
-        if end_idx > self.buffer_size:
-            # Wrap around the buffer end
-            upper_len = self.buffer_size - self.step
-            self.states[self.step : self.buffer_size] = states[:upper_len]
-            self.next_states[self.step : self.buffer_size] = next_states[:upper_len]
-            self.states[: end_idx - self.buffer_size] = states[upper_len:]
-            self.next_states[: end_idx - self.buffer_size] = next_states[upper_len:]
+        batch_size = states.shape[0]
+        end = self.step + batch_size
+
+        if end <= self.buffer_size:
+            self.states[self.step : end] = states
+            self.next_states[self.step : end] = next_states
         else:
-            self.states[start_idx:end_idx] = states
-            self.next_states[start_idx:end_idx] = next_states
+            # Wrap around
+            first_part = self.buffer_size - self.step
+            self.states[self.step :] = states[:first_part]
+            self.next_states[self.step :] = next_states[:first_part]
+            remainder = batch_size - first_part
+            self.states[:remainder] = states[first_part:]
+            self.next_states[:remainder] = next_states[first_part:]
 
-        self.num_samples = min(self.buffer_size, max(end_idx, self.num_samples))
-        self.step = (self.step + num_states) % self.buffer_size
+        # Update pointers
+        self.step = end % self.buffer_size
+        self.num_samples = min(self.buffer_size, self.num_samples + batch_size)
 
     def feed_forward_generator(
-        self, num_mini_batch: int, mini_batch_size: int
+        self,
+        num_mini_batch: int,
+        mini_batch_size: int,
     ) -> Generator[Tuple[torch.Tensor, torch.Tensor], None, None]:
         """
         Yield mini-batches of (state, next_state) tuples from the buffer.
@@ -78,11 +98,20 @@ class ReplayBuffer:
         Yields:
             Tuple[Tensor, Tensor]: A mini-batch of states and next_states.
         """
-        for _ in range(num_mini_batch):
-            sample_idxs = np.random.choice(
-                self.num_samples, size=mini_batch_size, replace=False
-            )
-            yield (
-                self.states[sample_idxs].to(self.device),
-                self.next_states[sample_idxs].to(self.device),
-            )
+        total = num_mini_batch * mini_batch_size
+        assert (
+            total <= self.num_samples
+        ), f"Not enough samples in buffer: requested {total}, but have {self.num_samples}"
+
+        # Generate a random permutation of valid indices on-device
+        indices = torch.randperm(self.num_samples, device=self.device)[:total]
+
+        for i in range(num_mini_batch):
+            batch_idx = indices[i * mini_batch_size : (i + 1) * mini_batch_size]
+            yield self.states[batch_idx], self.next_states[batch_idx]
+
+    def __len__(self) -> int:
+        """
+        Return the number of valid samples currently stored in the buffer.
+        """
+        return self.num_samples