This pull request improves the `ReplayBuffer` class for better performance, and introduces a benchmarking script to evaluate its efficiency. The buffer now accepts both `str` and `torch.device` for device selection, and all operations are performed directly on the specified device.

The main updates include optimized batched insertion with proper device handling, and a faster sampling implementation using `torch.randperm` instead of NumPy. The `feed_forward_generator` now includes validation to ensure enough samples are available, and sampling is fully on-device for better performance.

A new benchmarking script measures throughput and timing for insertion, sampling, and combined operations. Results demonstrate significant performance gains:

**Before (CUDA):**  
- **Insert:** 204,800 samples in **1.6 ms** → **127.5M samples/s**  
- **Sample:** 20,480 samples in **85.2 ms** → **240K samples/s**  
- **Combined:** 256,000 ops in **157.8 ms** → **1.62M ops/s**

**After (CUDA):**  
- **Insert:** 204,800 samples in **1.6 ms** → **128.0M samples/s**  
- **Sample:** 20,480 samples in **1.1 ms** → **19.3M samples/s**  
- **Combined:** 256,000 ops in **25.9 ms** → **9.88M ops/s**