@GiulioRomualdi 2025-05-05
The only impact is the **stricter removal** of duplicates across mini‑batches and the new assertion that prevents sampling more items than are available.